==> elevator-downtime/notebook/1-cleaning-geocode.ipynb <==

# Cleaning & Geolocating

## Setup Python and R environment
you can ignore this section

%load_ext rpy2.ipython
%load_ext autoreload
%autoreload 2

%matplotlib inline  
from matplotlib import rcParams
rcParams['figure.figsize'] = (16, 100)

import warnings
from rpy2.rinterface import RRuntimeWarning
warnings.filterwarnings("ignore") # Ignore all warnings
# warnings.filterwarnings("ignore", category=RRuntimeWarning) # Show some warnings

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from IPython.display import display, HTML

%%javascript
// Disable auto-scrolling
IPython.OutputArea.prototype._should_scroll = function(lines) {
    return false;
}

%%R

# My commonly used R imports

require('tidyverse')

##  Load data

df = pd.read_csv('data/input/2023.csv')
pd.set_option('display.max_colwidth', None)
df.sample(5)

%%R
df <- read_csv('data/input/2023.csv')

## Clean data

# filter down to elevators only
df_ele = df[df['Equipment Description'].str.startswith('ELE')]

# create column `station MRN` to later merge with `Station ID` in another file to geocode
df_ele['Station MRN'] = df_ele['Equipment Description'].str.extract(r'ELE: EL\d{2,3}\w? - (\d{2,3})\s-')

# check what the are the missing station MRN
# missing values are all elevators that are either located in New Utrecht Av (MRN: 73) or Dyckman St Quarters(MRN: TBD)

df_ele[df_ele['Station MRN'].isna()].to_csv('data/missing-value/2023_null_station_mrn.csv', index=False)
df_ele = df_ele.dropna(subset=['Station MRN'])

# convert string to float
df_ele['Station MRN'] = df_ele['Station MRN'].astype('Int64')

# stations with MRN > 523 are not regular subway stations
# filter down to subway elevators 
df_subway_ele = df_ele[df_ele['Station MRN'] <= 523]
print(df_subway_ele.shape)

# save to a new file
df_subway_ele.to_csv('data/intermediary/2023_ele.csv', index=False)
df_ele[df_ele['Station MRN'] > 523].to_csv('data/missing-value/2023_ele_not_subway.csv', index=False)

## Geocoding: Adding lat long to dataset

Merging datasets by `Station ID`(subway_station.csv) and `Station MRN` (my dataset)


# load the geostation file
df_geostation = pd.read_csv('data/input/subway_stations.csv')

# merge the geostation file with the df_ele
df_ele_merged = df_subway_ele.merge(df_geostation, left_on='Station MRN', right_on='Station ID', how='inner')
df_ele_merged = df_ele_merged.drop_duplicates(subset=['Outage'])

# rename the columns
df_ele_merged = df_ele_merged.rename(columns={'GTFS Latitude': 'lat',  'GTFS Longitude': 'long'})
df_ele_merged.to_csv('data/intermediary/2023_ele_geocode.csv', index=False)


## Convert lat/long to census geography codes 

(like 'GEOID', 'STATE', 'COUNTY', 'TRACT', 'BLOCK', etc...)

Same note as above, see [census-examples](https://github.com/data4news/census-examples) repository for examples or ask in the class slack channel if stuck.

import censusgeocode as cg
from concurrent.futures import ThreadPoolExecutor
from tqdm.notebook import tqdm

import requests_cache
cache = requests_cache.CachedSession("geocode_cache", backend="filesystem")

def geocode(lat, lng):
    try:
        url = "https://geocoding.geo.census.gov/geocoder/geographies/coordinates"
        params = {
            "x": lng,
            "y": lat,
            "benchmark": "Public_AR_Census2020",
            "vintage": "Census2020_Census2020",
            "format": "json"
        }
        response = cache.get(url, params=params)
        response.raise_for_status()
        data = response.json()
        census = data['result']['geographies']['Census Blocks'][0]
        return census
    except Exception as e:
        print(f"Error geocoding ({lat}, {lng}): {e}")
        return None

def bulk_geocode(latitudes, longitudes):
    """
    Geocode a list of latitudes and longitudes in parallel (for speed).
    """

    with ThreadPoolExecutor() as tpe:
        latitudes = df_ele_merged['lat']
        longitudes = df_ele_merged['long']
        mapped_results = tpe.map(geocode, latitudes, longitudes)
        data = list(tqdm(mapped_results, total=len(df_ele_merged)))

    return pd.DataFrame(data)

census_geos_df = bulk_geocode(df_ele_merged['lat'], df_ele_merged['long']) 
census_geos_df.head()

to_keep = ['GEOID', 'STATE', 'COUNTY', 'TRACT', 'BLOCK']
census_geos_df = census_geos_df[to_keep]
census_geos_df

## Output Data

Output your dataframe containing your data and the Census connector codes (like tract, block, etc...).

df_with_geos = pd.concat(
    [ 
        df_ele_merged.reset_index(drop=True),
        census_geos_df.reset_index(drop=True)
    ], 
    axis=1)

print(df_with_geos.shape)
df_with_geos.head()


df_with_geos.to_csv('data/intermediary/2023_subway_censusgeo.csv', index=False)
