==> elevator-downtime/notebook/3-t-test.ipynb <==
## 1. Setup

%load_ext rpy2.ipython
%load_ext autoreload
%autoreload 2

%matplotlib inline  
from matplotlib import rcParams
rcParams['figure.figsize'] = (16, 100)

import warnings
from rpy2.rinterface import RRuntimeWarning
warnings.filterwarnings("ignore") # Ignore all warnings
# warnings.filterwarnings("ignore", category=RRuntimeWarning) # Show some warnings

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from IPython.display import display, HTML
from datetime import datetime

%%javascript
// Disable auto-scrolling
IPython.OutputArea.prototype._should_scroll = function(lines) {
    return false;
}

%%R

# My commonly used R imports
require('flexplot')
require('tidyverse')
library(ggrepel)

## 2. Load census variable file

%%R -o df
df <- read_csv('data/intermediary/2023_subway_censusvar.csv')

df.shape

## 3. Creating column for 'downtime' & 'external' elevators

%%R -o df

# create column 'external' for rows in `Common Name` contains `X`
df <- df %>%
    mutate(external = grepl('X', `Common Name`)) 

# create new column 'downtime' 
# calculate the time between 'Out of Service Date' and 'Actual Return to Service Date'

df['Out of Service Date'] = pd.to_datetime(df['Out of Service Date'])
df['Actual Return to Service Date'] = pd.to_datetime(df['Actual Return to Service Date'])

df['downtime'] = df['Actual Return to Service Date'] - df['Out of Service Date'] 

# df['maintenance_time'] into hours
df['downtime'] = df['downtime'].dt.total_seconds() / 3600

# rename 'Outage Code' to 'outage_code'
df.rename(columns={'Outage Code': 'outage_code'}, inplace=True)
df.to_csv('data/output/2023_subway_downtime-external.csv', index=False)

# missing values to be dealt with
df[df['downtime'].isna()].to_csv('data/missing-value/2023_null_downtime.csv', index=False)

# check how many of the elevators are externally managed
df['external'].value_counts()

# downtime summary stat
df['downtime'].describe()

## 4. t-test for downtime between MTA and external elevators

%%R -i df

t.test(df %>% filter(external == 1) %>% select(downtime),
       df %>% filter(external == 0) %>% select(downtime))
